% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_city_id.R
\name{get_city_id}
\alias{get_city_id}
\title{get_city_id
Given the name of a Swedish municipality, it returns the corresponding id in the Kolada API.}
\usage{
get_city_id(city)
}
\arguments{
\item{city}{a string}
}
\value{
a string corresponding to the id
}
\description{
get_city_id
Given the name of a Swedish municipality, it returns the corresponding id in the Kolada API.
}
\examples{
get_city_id("Kiruna")
}
