% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get_data
Given a kpi id, year and city id it returns the corresponding data.}
\usage{
get_data(kpi_id = "", year = "", city_id = "")
}
\arguments{
\item{kpi_id}{a string}

\item{year}{a string}

\item{city_id}{a string}
}
\value{
a data frame
}
\description{
get_data
Given a kpi id, year and city id it returns the corresponding data.
}
\examples{
get_data(kpi_id = "N00403", year = "2019,2020", city_id = "0580")
}
