% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kpi_desc.R
\name{get_kpi_desc}
\alias{get_kpi_desc}
\title{get_kpi_desc
Given a kpi title it returns a short description related to the kpi}
\usage{
get_kpi_desc(kpi)
}
\arguments{
\item{kpi}{a string}
}
\value{
a string
}
\description{
get_kpi_desc
Given a kpi title it returns a short description related to the kpi
}
\examples{
get_kpi_desc("Ekologiskt odlad \u00e5kermark, andel (\%)")
}
