% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kpi_id.R
\name{get_kpi_id}
\alias{get_kpi_id}
\title{get_kpi_id
Given a kpi title it returns the corresponding id}
\usage{
get_kpi_id(kpi)
}
\arguments{
\item{kpi}{a string}
}
\value{
a string
}
\description{
get_kpi_id
Given a kpi title it returns the corresponding id
}
\examples{
get_kpi_id("Ekologiskt odlad \u00e5kermark, andel (\%)")
}
