% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kpi_title.R
\name{get_kpi_title}
\alias{get_kpi_title}
\title{get_kpi_title
Given a kpi id it returns the corresponding title}
\usage{
get_kpi_title(kpi_id)
}
\arguments{
\item{kpi_id}{a string}
}
\value{
a string
}
\description{
get_kpi_title
Given a kpi id it returns the corresponding title
}
\examples{
get_kpi_title("N00403")
}
