% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kolada_api.R
\name{kolada_api}
\alias{kolada_api}
\title{S3 class wrapper for Kolada API
Given a url it returns an S3 object of class kolada_api which wraps important results of the API call.}
\usage{
kolada_api(url)
}
\arguments{
\item{url}{a string representing a call to the Kolada api}
}
\value{
an S3 object of class kolada_api
}
\description{
S3 class wrapper for Kolada API
Given a url it returns an S3 object of class kolada_api which wraps important results of the API call.
}
\examples{
k_api <- kolada_api("http://api.kolada.se/v2/municipality?title=ale")
}
