% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_kolada_api.R
\name{print.kolada_api}
\alias{print.kolada_api}
\title{print.kolada_api
Print method for s3 objects of class kolada_api}
\usage{
\method{print}{kolada_api}(x, ...)
}
\arguments{
\item{x}{an object of class kolada_api}

\item{...}{other possible parameters}
}
\value{
a string which gets printed out
}
\description{
print.kolada_api
Print method for s3 objects of class kolada_api
}
\examples{
k_api <- kolada_api("http://api.kolada.se/v2/municipality?title=ale")
print(k_api)
}
